+++ Sources/Container/Async/AsyncContainer.swift
@@ -57,18 +57,47 @@ public actor AsyncContainer: AsyncDependencyResolving, AsyncDependencyRegisterin
     /// Register a dependency with an argument
     ///
     /// The argument is typically a parameter in an initiliazer of the dependency that is not registered in the same container,
-    /// therefore, it needs to be passed in `resolve` call
-    ///
-    /// DISCUSSION: This registration method doesn't have any scope parameter for a reason.
-    /// The container should always return a new instance for dependencies with arguments as the behaviour for resolving shared instances with arguments is undefined.
-    /// Should the argument conform to ``Equatable`` to compare the arguments to tell whether a shared instance with a given argument was already resolved?
-    /// Shared instances are typically not dependent on variable input parameters by definition.
-    /// If you need to support this usecase, please, keep references to the variable singletons outside of the container.
+    /// therefore, it needs to be passed in `resolve` call. This registration method doesn't have any scope parameter for a reason - the container
+    /// should always return a new instance for dependencies with arguments.
     ///
     /// - Parameters:
     ///   - type: Type of the dependency to register
     ///   - factory: Closure that is called when the dependency is being resolved
-    public func register<Dependency, Argument>(type: Dependency.Type, factory: @escaping FactoryWithArgument<Dependency, Argument>) async {
+    public func register<Dependency, Argument>(type: Dependency.Type, factory: @escaping FactoryWithOneArgument<Dependency, Argument>) async {
+        let registration = AsyncRegistration(type: type, scope: .new, factory: factory)
+
+        registrations[registration.identifier] = registration
+    }
+
+    // MARK: Register dependency with two arguments
+
+    /// Register a dependency with two arguments
+    ///
+    /// The arguments are typically parameters in an initializer of the dependency that are not registered in the same container,
+    /// therefore, they need to be passed in `resolve` call. This registration method doesn't have any scope parameter for a reason - the container
+    /// should always return a new instance for dependencies with arguments.
+    /// 
+    /// - Parameters:
+    ///   - type: Type of the dependency to register
+    ///   - factory: Closure that is called when the dependency is being resolved
+    public func register<Dependency, Argument1, Argument2>(type: Dependency.Type, factory: @escaping FactoryWithTwoArguments<Dependency, Argument1, Argument2>) async {
+        let registration = AsyncRegistration(type: type, scope: .new, factory: factory)
+
+        registrations[registration.identifier] = registration
+    }
+
+    // MARK: Register dependency with three arguments
+
+    /// Register a dependency with three arguments
+    ///
+    /// The arguments are typically parameters in an initializer of the dependency that are not registered in the same container,
+    /// therefore, they need to be passed in `resolve` call. This registration method doesn't have any scope parameter for a reason - the container
+    /// should always return a new instance for dependencies with arguments.
+    ///
+    /// - Parameters:
+    ///   - type: Type of the dependency to register
+    ///   - factory: Closure that is called when the dependency is being resolved
+    public func register<Dependency, Argument1, Argument2, Argument3>(type: Dependency.Type, factory: @escaping FactoryWithThreeArguments<Dependency, Argument1, Argument2, Argument3>) async {
         let registration = AsyncRegistration(type: type, scope: .new, factory: factory)
 
         registrations[registration.identifier] = registration
@@ -106,21 +135,60 @@ public actor AsyncContainer: AsyncDependencyResolving, AsyncDependencyRegisterin
 
         return try await getDependency(from: registration) as Dependency
     }
+
+    /// Resolve a dependency that was previously registered with `register` method
+    ///
+    /// If a dependency of the given type with the given arguments wasn't registered before this method call
+    /// the method throws ``ResolutionError.dependencyNotRegistered``
+    ///
+    /// - Parameters:
+    ///   - type: Type of the dependency that should be resolved
+    ///   - arguments: Aarguments that will passed as an input parameter to the factory method that was defined with `register` method
+    public func tryResolve<Dependency: Sendable, Argument1: Sendable, Argument2: Sendable>(type: Dependency.Type, argument1: Argument1, argument2: Argument2) async throws -> Dependency {
+        let identifier = RegistrationIdentifier(type: type, argument1: Argument1.self, argument2: Argument2.self)
+
+        let registration = try getRegistration(with: identifier)
+
+        return try await getDependency(from: registration, with: (argument1, argument2)) as Dependency
+    }
+
+    /// Resolve a dependency that was previously registered with `register` method
+    ///
+    /// If a dependency of the given type with the given arguments wasn't registered before this method call
+    /// the method throws ``ResolutionError.dependencyNotRegistered``
+    ///
+    /// - Parameters:
+    ///   - type: Type of the dependency that should be resolved
+    ///   - arguments: Aarguments that will passed as an input parameter to the factory method that was defined with `register` method
+    public func tryResolve<Dependency: Sendable, Argument1: Sendable, Argument2: Sendable, Argument3: Sendable>(type: Dependency.Type, argument1: Argument1, argument2: Argument2, argument3: Argument3) async throws -> Dependency {
+        let identifier = RegistrationIdentifier(type: type, argument1: Argument1.self, argument2: Argument2.self, argument3: Argument3.self)
+
+        let registration = try getRegistration(with: identifier)
+
+        return try await getDependency(from: registration, with: (argument1, argument2, argument3)) as Dependency
+    }
 }
 
 // MARK: Private methods
 private extension AsyncContainer {
     func getRegistration(with identifier: RegistrationIdentifier) throws -> AsyncRegistration {
-        guard let registration = registrations[identifier] else {
-            throw ResolutionError.dependencyNotRegistered(
-                message: "Dependency of type \(identifier.description) wasn't registered in container \(self)"
+        if let registration = registrations[identifier] {
+            return registration
+        }
+
+        if let matchingIdentifier = registrations.keys.first(where: { $0.typeIdentifier == identifier.typeIdentifier }) {
+            throw ResolutionError.unmatchingArgumentType(
+                message: "Registration of type \(matchingIdentifier.description) doesn't accept arguments of type \(identifier.description)"
             )
         }
 
-        return registration
+        throw ResolutionError.dependencyNotRegistered(
+            message: "Dependency of type \(identifier.description) wasn't registered in container \(self)"
+        )
+
     }
 
-    func getDependency<Dependency: Sendable>(from registration: AsyncRegistration, with argument: (any Sendable)? = nil) async throws -> Dependency {
+    func getDependency<Dependency: Sendable>(from registration: AsyncRegistration, with argument: Any? = nil) async throws -> Dependency {
         switch registration.scope {
         case .shared:
             if let dependency = sharedInstances[registration.identifier] as? Dependency {