+++ Sources/Models/Async/AsyncRegistration.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-typealias AsyncRegistrationFactory = @Sendable (any AsyncDependencyResolving, (any Sendable)?) async throws -> any Sendable
+typealias AsyncRegistrationFactory = @Sendable (any AsyncDependencyResolving, Any?) async throws -> any Sendable
 
 /// Object that represents a registered dependency and stores a closure, i.e. a factory that returns the desired dependency
 struct AsyncRegistration: Sendable {
@@ -34,6 +34,36 @@ struct AsyncRegistration: Sendable {
             }
 
             return await factory(resolver, argument)
+        }
+    }
+
+    /// Initializer for registrations that expect two variable arguments passed to the factory closure when the dependency is being resolved
+    init<T: Sendable, Argument1: Sendable, Argument2: Sendable>(type: T.Type, scope: DependencyScope, factory: @Sendable @escaping (any AsyncDependencyResolving, Argument1, Argument2) async -> T) {
+        let registrationIdentifier = RegistrationIdentifier(type: type, argument1: Argument1.self, argument2: Argument2.self)
+
+        identifier = registrationIdentifier
+        self.scope = scope
+        asyncRegistrationFactory = { resolver, arg in
+            guard let arguments = arg as? (Argument1, Argument2) else {
+                throw ResolutionError.unmatchingArgumentType(message: "Registration of type \(registrationIdentifier.description) doesn't accept arguments of type (\(Argument1.self), \(Argument2.self))")
+            }
+
+            return await factory(resolver, arguments.0, arguments.1)
+        }
+    }
+
+    /// Initializer for registrations that expect three variable arguments passed to the factory closure when the dependency is being resolved
+    init<T: Sendable, Argument1: Sendable, Argument2: Sendable, Argument3: Sendable>(type: T.Type, scope: DependencyScope, factory: @Sendable @escaping (any AsyncDependencyResolving, Argument1, Argument2, Argument3) async -> T) {
+        let registrationIdentifier = RegistrationIdentifier(type: type, argument1: Argument1.self, argument2: Argument2.self, argument3: Argument3.self)
+
+        identifier = registrationIdentifier
+        self.scope = scope
+        asyncRegistrationFactory = { resolver, arg in
+            guard let arguments = arg as? (Argument1, Argument2, Argument3) else {
+                throw ResolutionError.unmatchingArgumentType(message: "Registration of type \(registrationIdentifier.description) doesn't accept arguments of type (\(Argument1.self), \(Argument2.self), \(Argument3.self))")
+            }
+
+            return await factory(resolver, arguments.0, arguments.1, arguments.2)
         }
     }
 }